Added the --hashInputSizeLimit flag.
And more importantly, that now the hash is calculated only on the first
'hashInputSizeLimit' bytes of the file (defaulting to 1 MiB), which is
probably sufficient for uniqueness in the vast majority of situations,
while saving an immense amount of disk I/O (and thus wallclock time)
for many uses (e.g. mine of reading ~25 MiB images).