Display an ASCII progress bar instead of an indeterminate-length stream of pyramids.
This includes a rudimentary time remaining estimation, which readily and regularly falls afoul of disk caching - i.e. the first few files are in the disk cache and compare very quickly, thus skewing the timings and making it produce Microsoft-time estimates for pretty much the rest of the operation.  A better implementation might be to use a moving window, which won't be able to anticipate changes in basic I/O speed like that, but will at least fairly quickly settle down to accurate estimates once the disk cached portion are compared.