Now disable caching & read-ahead for files when hashing them, on the assumption that most files are unique (and hash collisions are rare) so we won't ever return to most of them to actually compare their contents byte-for-byte.  This should help avoid undue pressure on the system when the files being compared just won't fit into memory (and other files from other user programs may be there usefully otherwise), and make repeated runs of this tool faster by maximising use of the file cache for files that are actually being compared byte-for-byte.
This entails opening the file manually and handing the file descriptor to libdispatch (as there's no way to get the file descriptor, or otherwise set fcntl attributes, on what libdispatch opens implicitly).  Previously it was implied and relied upon that a dispatch_io_t wouldn't actually use an underlying file descriptor until the first read request, and so any number of them could be "opened" simultaneously without creating resource exhaustion.  Since that's not the case when opening the file descriptors manually, some refactoring was necessary to move the actual open behind the concurrency limiting semaphore.