Concurrency limits when doing the final, full file comparison are now based on whether both files are on SSDs or not - if not, the concurrency limit is set to 1 (i.e. no concurrency) on the assumption that the files are non-trivial in size and not significantly fragmented, therefore it's best not to burden the spinning rust with multiple reads simultaneously as it will probably time-share amongst them and thus waste a lot of time seeking between them.
SSDs get up to 64x concurrency (remember that each file comparison is two files to read from, so it's potentially 2x that if both files are on the same SSD), chosen pretty arbitrarily just because in the presumed typical case where the two files are on different disks, the concurrency would then be 32x and that's the SATA NCQ size limit.  PCIe or NVMe SSDs perhaps need even greater concurrency, though presumably there's diminishing return on today's hardware which has *at most* 36 logical cores (18-core iMac Pro).

Technically this means, in the presence of some comparisons between files both on the same SSD and some with only one file on that same SSD and the other on spinning rust, that there can be 64 + 1 = 65 outstanding reads to an SSD, but the incremental difference in concurrency is very likely insignificant.