Now wait until the I/O stream is _actually_ closed, when verifying duplicity of files, to signal the concurrency limiter.
This ensures that the file descriptors underlying the completed comparison are actually closed & released before another comparison can take it's slot, thus ensuring that (for reasonable concurrency limits) the open file descriptor limit is not hit.

Alas dispatch_semaphore_wait() can't decrement by anything but 1, so there's a bit of awkward extra complexity in compareFiles() in order to make this happen, but it's quite functional.  (I did try the obviously naive approach of just calling dispatch_semaphore_wait() twice, but of course that inevitably leads to deadlock when all available concurrency slots end up blocked holding the first unit off the semaphore, but unable to get a second)